Ask travis to run simulation and tutorial loadtest separately.
@leplatrem review.
