+++ .travis.yml
@@ -9,14 +9,16 @@ env:
     - TOX_ENV=pypy
     - TOX_ENV=flake8
     - TOX_ENV=docs
-    - ACTION=loadtest
+    - ACTION=loadtest_tutorial
+    - ACTION=loadtest_simulation
 before_script:
     - psql -c "CREATE DATABASE testdb ENCODING 'UTF8' TEMPLATE template0;" -U postgres
 install:
-    - if [[ $ACTION != loadtest ]]; then pip install tox; fi
+    - if [[ $ACTION != loadtest_tutorial && $ACTION != loadtest_simulation ]]; then pip install tox; fi
 script:
-    - if [[ $ACTION != loadtest ]]; then tox -e $TOX_ENV; fi
-    - if [[ $ACTION == loadtest ]]; then make loadtest-check; fi
+    - if [[ $ACTION != loadtest_tutorial && $ACTION != loadtest_simulation ]]; then tox -e $TOX_ENV; fi
+    - if [[ $ACTION == loadtest_simulation ]]; then make loadtest-check-simulation; fi
+    - if [[ $ACTION == loadtest_tutorial ]]; then make loadtest-check-tutorial; fi
 after_success:
     # Report coverage results to coveralls.io
     - pip install coveralls
